refactor : 이벤트 관련 리팩토링
