+++ itcook-domain/src/main/java/com/itcook/cooking/domain/domains/user/service/AuthCodeRedisService.java
@@ -5,7 +5,7 @@ import static com.itcook.cooking.domain.infra.email.EmailTemplate.AUTH_EMAIL;
 import com.itcook.cooking.domain.common.errorcode.UserErrorCode;
 import com.itcook.cooking.domain.common.exception.ApiException;
 import com.itcook.cooking.domain.common.utils.RandomCodeUtils;
-import com.itcook.cooking.domain.infra.email.EmailSendEvent;
+import com.itcook.cooking.domain.common.events.email.EmailSendEvent;
 import com.itcook.cooking.domain.infra.redis.RedisService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
