+++ itcook-domain/src/main/java/com/itcook/cooking/domain/domains/user/service/UserDomainService.java
@@ -109,7 +109,6 @@ public class UserDomainService {
     public void leaveUser(String email) {
         ItCookUser user = findUserByEmail(email);
         user.delete();
-//        userAdaptor.saveUser(user); // 이벤트 발생을 위한 save 호출
     }
