--- itcook-domain/src/main/java/com/itcook/cooking/domain/domains/user/service/UserScheduler.java
@@ -1,16 +1,0 @@
-package com.itcook.cooking.domain.domains.user.service;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Service;
-
-@Slf4j
-@Service
-public class UserScheduler {
-
-
-//    @Scheduled(cron = "0 * * * * *")
-    public void update() {
-
-    }
-}
