+++ itcook-domain/src/main/java/com/itcook/cooking/domain/infra/email/AuthCodeService.java
@@ -1,5 +1,7 @@
 package com.itcook.cooking.domain.infra.email;
 
+import com.itcook.cooking.domain.common.events.email.EmailSendEvent;
+
 public interface AuthCodeService {
 
     void sentAuthCode(EmailSendEvent emailSendEvent);
