+++ itcook-domain/src/test/java/com/itcook/cooking/domain/domains/mock/MockEmailService.java
@@ -1,7 +1,7 @@
 package com.itcook.cooking.domain.domains.mock;
 
 import com.itcook.cooking.domain.infra.email.AuthCodeService;
-import com.itcook.cooking.domain.infra.email.EmailSendEvent;
+import com.itcook.cooking.domain.common.events.email.EmailSendEvent;
 
 public class MockEmailService implements AuthCodeService {
