+++ itcook-infra/src/test/java/com/itcook/cooking/infra/mock/FakeEmailSendService.java
@@ -1,7 +1,7 @@
 package com.itcook.cooking.infra.mock;
 
 import com.itcook.cooking.domain.infra.email.AuthCodeService;
-import com.itcook.cooking.domain.infra.email.EmailSendEvent;
+import com.itcook.cooking.domain.common.events.email.EmailSendEvent;
 
 public class FakeEmailSendService implements AuthCodeService {
